+++ crates/engine/src/plugin_manager.rs
@@ -1,14 +1,18 @@
 use crate::error::PluginManagerError as Error;
-use crate::plugin::Plugin;
+use crate::plugin::{Plugin, WasmPlugin};
 
 pub(crate) type Result<T> = std::result::Result<T, Error>;
 
-/// The `PluginManager` trait allows an object to manage a set of plugins.
+pub type DefaultPluginManager = PluginManager<WasmPlugin>;
+
+/// The `PluginHandler` trait allows an object to manage a set of plugins.
 pub trait PluginHandler {
+    fn new() -> Self;
+
     /// Run all registered plugins.
     fn run_plugins(&mut self) -> Result<()>;
 
-    /// Register a new plugin for this manager to manage.
+    /// Register a new plugin to handle.
     fn register_plugin(&mut self, path: &str) -> Result<()>;
 }
 
@@ -19,15 +23,22 @@ pub trait PluginHandler {
 /// - Loading new Wasm-based plugins.
 /// - Running plugins when requested.
 #[derive(Default)]
-pub struct PluginManager {
+pub struct PluginManager<T: Plugin> {
     /// The list of plugins this plugin manager is responsible for.
-    plugins: Vec<Plugin>,
+    plugins: Vec<T>,
 
-    // The wasm cache used by `wasmtime`.
+    // The wasm cache used by the `wasmtime` Wasm runtime.
     plugin_store: wasmtime::Store,
 }
 
-impl PluginHandler for PluginManager {
+impl<T: Plugin> PluginHandler for PluginManager<T> {
+    fn new() -> Self {
+        PluginManager {
+            plugins: vec![],
+            plugin_store: wasmtime::Store::default(),
+        }
+    }
+
     fn run_plugins(&mut self) -> Result<()> {
         for plugin in &self.plugins {
             plugin.run()?
@@ -41,7 +52,7 @@ impl PluginHandler for PluginManager {
 
         let module = Module::from_file(&self.plugin_store, path).map_err(|err| (path, err))?;
         let instance = Instance::new(&module, &[]).map_err(|err| (path, err))?;
-        let plugin = Plugin::new(instance);
+        let plugin = T::new(instance);
 
         self.plugins.push(plugin);
 
@@ -52,6 +63,7 @@ impl PluginHandler for PluginManager {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::plugin::WasmPlugin;
     use tempfile::NamedTempFile;
 
     mod run_plugins {
@@ -59,14 +71,14 @@ mod tests {
 
         #[test]
         fn empty() {
-            let mut manager = PluginManager::default();
+            let mut manager = manager();
 
             assert!(manager.run_plugins().is_ok())
         }
 
         #[test]
         fn multiple() {
-            let mut manager = PluginManager::default();
+            let mut manager = manager();
 
             let p = plugin(r#"(module (func (export "_run")))"#);
             manager.plugins.push(p);
@@ -79,7 +91,7 @@ mod tests {
 
         #[test]
         fn with_failure() {
-            let mut manager = PluginManager::default();
+            let mut manager = manager();
 
             let p = plugin(r#"(module (func (export "_run")))"#);
             manager.plugins.push(p);
@@ -100,14 +112,14 @@ mod tests {
         fn valid() {
             let (_guard, path) = wasm(r#"(module (func (export "_run")))"#);
 
-            assert!(PluginManager::default().register_plugin(&path).is_ok())
+            assert!(manager().register_plugin(&path).is_ok())
         }
 
         #[test]
         fn invalid_wasm() {
             let (_guard, path) = wasm(r#"INVALID"#);
 
-            let err = PluginManager::default().register_plugin(&path).unwrap_err();
+            let err = manager().register_plugin(&path).unwrap_err();
 
             assert_eq!(err.to_string(), format!("invalid wasm module `{}`", path))
         }
@@ -116,13 +128,17 @@ mod tests {
         fn missing_file() {
             let path = "/missing/file";
 
-            let err = PluginManager::default().register_plugin(&path).unwrap_err();
+            let err = manager().register_plugin(&path).unwrap_err();
 
             assert_eq!(
                 err.to_string(),
                 format!("inaccessible wasm module `{}` (NotFound)", path)
             )
         }
+    }
+
+    fn manager() -> PluginManager<WasmPlugin> {
+        PluginManager::new()
     }
 
     fn wasm(wasm: &str) -> (NamedTempFile, String) {
@@ -138,13 +154,13 @@ mod tests {
         (file, path)
     }
 
-    fn plugin(wasm: &str) -> Plugin {
+    fn plugin(wasm: &str) -> WasmPlugin {
         use wasmtime::{Instance, Module};
 
         let store = wasmtime::Store::default();
         let module = Module::new(&store, wasm).unwrap();
         let instance = Instance::new(&module, &[]).unwrap();
 
-        Plugin::new(instance)
+        WasmPlugin::new(instance)
     }
 }